requirements.txt
click~=7.1.2
pyautogui~=0.9.52
Pillow~=8.1.0